+++ src/plotgrid.jl
@@ -1,17 +1,18 @@
 """
-    plotscore(x, y; resolution = (500, 350), step = 5, 
+    plotgrid(indx, r; resolution = (500, 350), step = 5, 
         color = nothing, kwargs...)
-    plotscore(x, y, group; resolution = (500, 350), step = 5, 
+    plotgrid(indx, r, group; resolution = (500, 350), step = 5, 
         color = nothing, kwargs...)
 
-Plotting model performances (scores) after validation.
-* `x` : A variable representing the parameter(s).
-* `y` : The model scores (e.g. error rates) for the values of `x`. 
-* `group` : Grouping parameters if multiple.
+Plot error or performance rates of model predictions.
+* `indx` : A variable representing the grid of model parameters, e.g. nb. LVs if PLSR models.
+* `r` : The error/performance rates for the values of `x`. 
+* `group` : Categorical variable defining groups. 
+    A separate line is plotted for each level of `group`.
 * 'resolution' : Resolution (horizontal, vertical) of the figure.
 * `color` : Set color. If `group` if used, must be a vector of same length
     as the number of levels in `group`.
-* `kwargs` : Optional arguments to pass in `Axis`.
+* `kwargs` : Optional arguments to pass in `Axis` of CairoMakie.
 
 ## Examples
 ```julia
@@ -34,7 +35,7 @@ ytest = rmrow(y, s)
 nlv = 0:20
 res = gridscorelv(Xtrain, ytrain, Xtest, ytest;
     score = rmsep, fun = plskern, nlv = nlv)
-plotscore(res.nlv, res.y1;
+plotgrid(res.nlv, res.y1;
     xlabel = "Nb. LVs", ylabel = "RMSEP").f
 
 nlvdis = 15 ; metric = ["mahal" ;]
@@ -44,32 +45,32 @@ nlv = 0:20
 res = gridscorelv(Xtrain, ytrain, Xtest, ytest;
     score = rmsep, fun = lwplsr, pars = pars, nlv = nlv)
 group = string.("h=", res.h, " k=", res.k)
-plotscore(res.nlv, res.y1, group;
+plotgrid(res.nlv, res.y1, group;
     xlabel = "Nb. LVs", ylabel = "RMSECV").f
 ```
 """ 
-function plotscore(x, y; resolution = (500, 350), step = 5, 
+function plotgrid(indx, r; resolution = (500, 350), step = 5, 
         color = nothing, kwargs...)
-    x = vec(x)
-    y = vec(y)
-    xticks = collect(minimum(x):step:maximum(x))
+    indx = collect(vec(indx))
+    r = vec(r)
+    xticks = collect(minimum(indx):step:maximum(indx))
     f = Figure(resolution = resolution)
     ax = Axis(f; xticks = (xticks, string.(xticks)), kwargs...)
     if isnothing(color)
-        lines!(ax, x, y)
+        lines!(ax, indx, r)
     else
-        lines!(ax, x, y; color = color)
+        lines!(ax, indx, r; color = color)
     end
     f[1, 1] = ax
     (f = f, ax = ax)
 end
 
-function plotscore(x, y, group; resolution = (700, 350), step = 5, 
+function plotgrid(indx, r, group; resolution = (700, 350), step = 5, 
         color = nothing, kwargs...)
-    x = vec(x)
-    y = vec(y)
+    indx = collect(vec(indx))
+    r = vec(r)
     group = vec(group)
-    xticks = collect(minimum(x):step:maximum(x))
+    xticks = collect(minimum(indx):step:maximum(indx))
     lev = sort(unique(group))
     nlev = length(lev)
     println(nlev)
@@ -77,13 +78,13 @@ function plotscore(x, y, group; resolution = (700, 350), step = 5,
     ax = Axis(f; xticks = (xticks, string.(xticks)), kwargs...)
     for i = 1:nlev
         s = group .== lev[i]
-        zx = x[s]
-        zy = y[s]
+        x = indx[s]
+        y = r[s]
         lab = string(lev[i])
         if isnothing(color)
-            lines!(ax, zx, zy; label = lab)
+            lines!(ax, x, y; label = lab)
         else
-            lines!(ax, zx, zy; label = lab, color = color[i])
+            lines!(ax, x, y; label = lab, color = color[i])
         end
     end
     f[1, 1] = ax