+++ src/wshenk.jl
@@ -21,7 +21,7 @@ least squares. Analytica Chimica Acta 504, 279–289. https://doi.org/10.1016/j.
 
 ## Examples 
 ```julia 
-using JLD2, CairoMakie
+using JLD2
 mypath = dirname(dirname(pathof(Jchemo)))
 db = joinpath(mypath, "data", "cassav.jld2") 
 @load db dat